Refactor: move favorite to moduledomain
