+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/favorite/core/FavoriteStore.java
@@ -1,6 +1,6 @@
 package kernel.jdon.moduleapi.domain.favorite.core;
 
-import kernel.jdon.favorite.domain.Favorite;
+import kernel.jdon.moduledomain.favorite.domain.Favorite;
 
 public interface FavoriteStore {
 	Favorite save(Favorite favorite);
