+++ module-api/src/main/java/kernel/jdon/moduleapi/domain/favorite/infrastructure/FavoriteRepository.java
@@ -5,8 +5,8 @@ import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 
-import kernel.jdon.favorite.domain.Favorite;
-import kernel.jdon.favorite.repository.FavoriteDomainRepository;
+import kernel.jdon.moduledomain.favorite.domain.Favorite;
+import kernel.jdon.moduledomain.favorite.repository.FavoriteDomainRepository;
 
 public interface FavoriteRepository extends FavoriteDomainRepository {
