+++ module-domain/src/main/java/kernel/jdon/moduledomain/favorite/repository/FavoriteDomainRepository.java
@@ -1,8 +1,8 @@
-package kernel.jdon.favorite.repository;
+package kernel.jdon.moduledomain.favorite.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kernel.jdon.favorite.domain.Favorite;
+import kernel.jdon.moduledomain.favorite.domain.Favorite;
 
 public interface FavoriteDomainRepository extends JpaRepository<Favorite, Long> {
 }
